+++ src/Core/Dfe.Complete.Infrastructure/Database/CompleteContext.cs
@@ -661,6 +661,7 @@ public partial class CompleteContext : DbContext
             .HasColumnName("updated_at");
         projectConfiguration.Property(e => e.NurseryArrangement).HasColumnName("nursery_arrangement");
         projectConfiguration.Property(e => e.ConfirmDBSChecks).HasColumnName("confirm_dbs_checks");
+        projectConfiguration.Property(e => e.CommercialTransferAgreementUnamended).HasColumnName("transfer_agreement_unamended");        
     }
 
     private static void ConfigureTransferTasksData(EntityTypeBuilder<TransferTasksData> projectConfiguration)