+++ src/Tests/Dfe.Complete.CypressTests/cypress/api/taskHelper.ts
@@ -53,6 +53,7 @@ export class TaskHelper {
             questionsReceived: false,
             questionsChecked: false,
             saved: false,
+            unamended: false,
         };
 
         switch (status) {
@@ -71,6 +72,7 @@ export class TaskHelper {
                     questionsReceived: true,
                     questionsChecked: true,
                     saved: true,
+                    unamended: true,
                 });
 
             default: