+++ CHANGELOG.md
@@ -38,6 +38,11 @@ See the [full commit history](https://github.com/DFE-Digital/complete-conversion
 
 ---
 
+### Changed
+- [279041](https://dev.azure.com/dfe-gov-uk/Academies-and-Free-Schools-SIP/_workitems/edit/279041) - Commercial Transfer Agreement
+  
+---
+
 ## [1.46.1](https://github.com/DFE-Digital/complete-conversions-transfers-changes/releases/tag/production-2026-05-21.1562) - 2026-05-21
 
 ### Changed